Added another file to store the main code
