+++ spec/bot_spec.rb
@@ -2,13 +2,14 @@
 
 # spec/bot_spec.rb
 
-require_relative '../lib/bot'
+require_relative '../lib/main'
 
-describe Calculator do
-  describe '#add' do
-    it 'returns the sum of two numbers' do
-      calculator = Calculator.new
-      expect(calculator.add(5, 2)).to eql(7)
+describe Answer do
+  describe '.ans_img' do
+  let(:ans_test) {Answer.new}
+    it 'The reply is the expected due the message' do
+      ans_test.ans_img('When did he leave?')
+      expect(ans_test.reply).to eql('../img/try_again.jpg')
     end
   end
 end
